Include cookbook versions in nodes report.
This change also shifts the rendering of output to the calling command, leaving the
nodes report to return an array of NodeReportItems.  This will let us
manage output/rendering of the data based on the options provided.

This is subject to further refinement as we determine the best shape the
output for computers and people.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>