Protect coverage update task from failing
This commit is adding a simple check to verify if the update coverage
task should be run or not, in the case where a PR doesn't modify the
code coverage and we don't add the tag to avoid the coverage update,
then this test would fail trying to modify something that hasn't
changed. By checking both values first we avoid failures and advice the
user to use the GH tag.

Signed-off-by: Salim Afiune <afiune@chef.io>