Fix Habitat builds
The core/scaffolding-go has a feature that, if a Go project has a
`Makefile` inside the source directory (`/src`), it will execute `make`
to build the go binary. Since I added a `Makefile`, the pipeline has
been failing for a couple days. This change detects if we are running in
our CI pipeline, and if so, changes the default make task to build the
go binary as the scaffolding does it.

Signed-off-by: Salim Afiune <afiune@chef.io>