Generate the chefstyle formatter at runtime.
This simplifies distribution - this formatter
is specialized enough that it doesn't make sense to add it
chefstyle itself; but keeping it as a .rb file means
customizing the chef-analyze omnibus and habitat builds
to include it in the package.

Including it in the source of the file seems a more sane
way of managing it.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>