(flag rename) skip-unused exclude unused cookbooks
to give better clarity to the flag I have renamed it to be
`--skip-unused` so that it is clear that when provided, the tool will
exclude the unused cookbooks. The default behavior is still to include
the unused cookbooks, the rename is to avoid turning it on by default.

Additionally, the flag is now being used.

Signed-off-by: Salim Afiune <afiune@chef.io>