Improve table formatting and width
This condenses the table formatting as much as possible,
and adds a warning footer when terminal size is smaller than
the minimum size we assume (120 col).

We may need to further adjust this; but better still might be to
take a closer look at the tablewriter package and seeing if we can
improve on its formatting.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>