Move `Array` function out of `Node` struct
The conversion logic was getting specialized to include formatting
details.  Moved it into the formatter package instead, in the table
formatter.

This also required making the function public in the format package,
since that allows us to use the same `_test` package naming conventions
that we follow elsewhere and still test the function.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>