Use actual table output size for warning message
This change stops trying to pre-allocate most of the column widths, and
allows the tablewriter package to handle that.

It puts a soft constraint of 120 characters/ncols on each column,
but this will only be respected if the content will wrap - so it's still
possible to go beyond 120 characters.

Finally, it checks the length of the actual output and updates the
'adjust your screen size' note to include the needed size based on
content, instead of a 'recommended minimum'

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>