Separation of concerns - text render in cmd
This updates the formatters to return only a formatted report
and a formatted error string (if any).  It leaves it up to the
caller to handle writing that output to file or screen.

This should make the outputs of the various formatters more testable
since we can now provide inputs and look directly at the outputs.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>