run-cookstyle:(flag) reduce cookbooks report
This commit is reducing the cookbooks report command to only generate a
small report about the number of nodes affected. When a user wants to
run cookstyle on them, they will have to provide the new flag
`--run-cookstyle`.

The motivation here is to run the command as fast as possible to start
generating reports and letting the user to expand and modify them to
include extra data.

Signed-off-by: Salim Afiune <afiune@chef.io>