Move remaining IO out of reporting pkg
This moves remaining IO (messages and progress)
out of hte reporting package and into the top-level
command, because the reporting package modules
exist only to assemble reports.

Instead, progress events are now reported back to the caller
via a channel, and caller can display any feedback to the operator
as needed.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>