cookstyle:(config) use --force-default-config
Doing a little bit of research on cookstyle/rubocop I found an option
that will help use ignore any `.rubocop.yml` that might conflict with
the current state of things.

This is a possible solution instead of deleting the file we will ignore
it so that we do not modify the customers data just yet. We might do
that in next iteration to fix things on cookbooks/data.

Signed-off-by: Salim Afiune <afiune@chef.io>