Adds "chef analyze capture"
This adds an initial `capture` command to `chef analyze`.

Capture accepts a node as an argument and creates a chef repo
that contains:

* node object
* required cookbooks
* roles
* enviroments

This scope of this change extends to created the chef repository,
but does not include configuring a node.

This is a WIP.  To be completed in this PR:
* [ ] cookbook directories cannot include the version name.
* [ ] remove files from repo we don't need. This includes kitchen.yml,
      since someone using TK in the repo would not get
      the experience they would expect.
* [ ] tests

To be completed in next sprint:

* customize the kitchen.yml to use chef-zero to serve the repository
from the host (allowing for iteration)
* refactor the shareable pieces that are in `reporting` so that
  we can move the `capture` functionality out of the `reporting`
  package.
* policyfile support

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>