Generate a kitchen.yml as part of capture
In order to test against the captured node, it's necessary
to generate a kitchen.yml configuration appropriate for the node.

This change adds the generation of a kitchen.yml
as the last step of node capture. Currently it looks
at node attributes to determine the type of image to create,
but this is limited to platform-name+platform-version (eg ubuntu-18.04)

This is suitable for our initial set of cases, but we will need to add
more flexibility in determining the right image name and the provider
(currently always vagrant), and will need to add prompting in cases
where we can't determine the correct provider/image.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>