Interactive prompting to resolve cookbook sources
This modifies 'capture' to interactively
prompt for the checkout location(s) of the captured
node's cookbooks; and replaces the original cookbooks
from the server with symlinks to the checkout location(s).

This will allow operators to begin iterating on the
converge-upgrade-fix cycle without worrying about
needing to re-sync their local changes in the node repository
with an upstream cookbook.

This will re-prompt for additional checkout paths until
all cookbooks are accounted for, or no further checkout
paths are provided.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>