Update for compatibility with recent cobra changes
See  https://github.com/spf13/cobra/pull/1068/files

The ErrSubCommandRequired type was removed and is no longer supported
by cobra.  To work around this, we add a RunE function to `rootCmd`
that invokes helps and exits without an error.

This differs from cobra's default behavior in  the command will
continue to exit with a zero exit code when no arguments are given;
cobra  would exit non-zero.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>