Extend integration coverage.
This adds integration coverage for capture help output
as well as cookbook and node reports.

because this is integration now things are added that don't have test
coverage, the new tests now compare the exact output of the help commands to the actual output.

This has the advantage of inducing failures when we add output text without
also updating the texts to account for new behaviors.  If this works well, we can extend
it to the other integration tests that ensure correctness of rendered texts.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>