Support use via as `chef report` and `chef capture`
This updates descripitions and usage text to reflect
the new entry point of 'chef report', 'chef capture' instead of
'chef-analyze'.

This includes an alias for `upload` so that it can be accessed from
the `chef` invocation `chef report upload` -> `chef-analyze upload`,
and render the correct usage text.

Text corrections have also been made for clarity and consistency with other
CLI tool conventions.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>