Add policyfile support to 'chef capture'
This adds chef-capture support for policy-managed nodes.

It works in the same way that it does for non-policy-managed nodes -
it will capture the node's state and dependencies (in this case - policy group/policy/cookbook
artifacts) from a Chef Infra Server, and saves it into a local 'captured node' repository
where it's served up on test-kitchen VM via chef-zero.

This differs from running policyfile tests in TK in that Kitchen itself won't attempt to
resolve or download any cookbook dependencies - it relies on chef capture having captured the
complete node state, including policy/artifacts applied  into the node repo before kitchen converge.

This functionality relies on a new provisioner for TK, "chef_zero_capture". This
provisioner ships with chef-cli, and is now available in 3.0.4+ of the chef-cli gem.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>