Add --with-data-bags option to capture
With this change, data bags will not be downloaded by
default when `capture` is run. Until we are no longer
downloading _all_ data bags, we don't want to
download the without expliitly being told to.

This adds a new CaptureOpts stuct, and a DownloadDataBags
option, and the corresponding CLI option "--with-data-bags"

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>